requirements.txt
Django==2.2.1
django-model-utils==3.1.2
djangorestframework==3.8.2
Faker==8.10.3
psycopg2-binary==2.8.4
python-dateutil==2.8.2
pytz==2019.1
six==1.16.0
sqlparse==0.3.0
text-unidecode==1.3
